# music_app_old/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "diskrot: Create",
    "description": "The API for this application",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/v1/status/{id}": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Return inference status",
        "description": "This make a request against the inference server to determine if the song has finished process.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get song status.",
            "schema": {
              "$ref": "#/definitions/SongStatus"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SongStatus": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "b7e6a5d9-5f42-4a8b-9a38-47ef2e2a8df1"
        },
        "title": {
          "type": "string",
          "example": "Banger #1"
        },
        "filename": {
          "type": "string",
          "example": "s3://bucket/folder/song.wav"
        },
        "dt_created": {
          "type": "string",
          "example": "2025-03-12T00:00:00+05:00"
        },
        "processing_status": {
          "type": "string",
          "example": "NEW"
        }
      }
    },
...
  }
}